Wee_05_All_Ass_done

+++ Programming Fundamentals/week_05/Assignment_05/51/Example.java
@@ -3,7 +3,7 @@ import java.util.*;
 class Example {
 	public static void main(String [] args) {
 		boolean status = true;
-		int num = 1;
+		long num = 1;
 		while(status){
 			int count = 0;
 			for(int i = 1 ; i<=20 ; i++){
